Changed code
